+++ yellowstone-grpc-proto/src/plugin/message.rs
@@ -387,7 +387,7 @@ impl MessageEntry {
             slot: info.slot,
             index: info.index,
             num_hashes: info.num_hashes,
-            hash: Hash::new(info.hash),
+            hash: Hash::new_from_array(<[u8; HASH_BYTES]>::try_from(info.hash).unwrap()),
             executed_transaction_count: info.executed_transaction_count,
             starting_transaction_index: info
                 .starting_transaction_index